pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yavai"
version = "0.1.0"
description = "End-to-End AI Platform - Data Management and MLOps Library"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

authors = [
    { name = "Ahmad Lutfi Bayu Aji", email = "bayuaji732@gmail.com" }
]

keywords = [
    "data-management",
    "mlops",
    "machine-learning",
    "experiment-tracking",
    "mlflow",
    "s3",
    "jdbc"
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "boto3>=1.26.0",
    "python-dotenv>=0.19.0",
    "furl>=2.1.3",
    "mlflow>=2.0.0",
    "plotly>=5.0.0",
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "requests>=2.25.0",
    "openpyxl>=3.0.0",
    "xlrd>=2.0.0",
    "librosa>=0.9.0",
    "pydub>=0.25.0",
    "opencv-python-headless>=4.5.0",
    "pillow-heif>=0.10.0",
    "Pillow>=9.0.0",
    "pyhive[hive-pure-sasl]>=0.6.5",
    "jaydebeapi>=1.2.3",
    "psycopg2-binary>=2.9.0",
    "paramiko>=2.10.0",
    "pyreadstat>=1.1.5",
    "ipython>=7.0.0",
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/bayuaji732/yavai-python"
Repository = "https://github.com/bayuaji732/yavai-python"
Issues = "https://github.com/bayuaji732/yavai-python/issues"
Documentation = "https://github.com/bayuaji732/yavai-python#readme"

[tool.setuptools]
packages = [
    "yavai",
    "yavai.connections",
    "yavai.datasets",
    "yavai.io",
    "yavai.tracking",
    "yavai.utils"
]

[tool.setuptools.package-data]
yavai = ["*.env", "py.typed"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".eggs"]